app/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
pyjwt
requests
python-dotenv

